chore: rename primitives/type_utils.py to primitives/utils.py
